+++ packages/url-loader/src/index.ts
@@ -1,5 +1,5 @@
 export { constructCloudinaryUrl, transformationPlugins } from './lib/cloudinary';
-export type { ConstructUrlProps, PluginOptionsResize, PluginOptions, PluginResults } from './lib/cloudinary';
+export type { ConstructUrlProps } from './lib/cloudinary';
 
 export { effects, position, primary, text } from './constants/qualifiers';
 
@@ -9,5 +9,5 @@ export type { VideoOptions } from './types/video';
 
 export type { AnalyticsOptions, CloudinaryAnalyticsOptions } from './types/analytics';
 export type { ConfigOptions, CloudinaryConfigurationOptions } from './types/config';
-export type { PluginSettings, PluginOverrides } from './types/plugins';
+export type { PluginSettings, PluginOptions, PluginResults } from './types/plugins';
 export type { Qualifier, QualiferConverters } from './types/qualifiers';