+++ packages/url-loader/tests/lib/cloudinary.spec.js
@@ -182,18 +182,20 @@ describe('Cloudinary', () => {
 
     describe('cropping, resizing', () => {
 
-      it('should create a Cloudinary URL with widthResize smaller than width', () => {
+      it('should create a Cloudinary URL with a width, height, and custom crop', () => {
         const cloudName = 'customtestcloud';
         const deliveryType = 'upload';
         const publicId = 'myimage';
         const width = 900;
-        const widthResize = 600;
+        const height = 900;
+        const crop = 'auto';
 
         const url = constructCloudinaryUrl({
           options: {
             src: publicId,
             width,
-            widthResize
+            height,
+            crop
           },
           config: {
             cloud: {
@@ -201,21 +203,23 @@ describe('Cloudinary', () => {
             }
           }
         });
-        expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/${deliveryType}/c_limit,w_${width}/c_limit,w_${widthResize}/f_auto/q_auto/${publicId}`);
+        expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/${deliveryType}/c_${crop},w_${width},h_${height},g_auto/f_auto/q_auto/${publicId}`);
       });
 
-      it('should create a Cloudinary URL with widthResize same as width', () => {
+      it('should create a Cloudinary URL with a aspect ratio, custom crop, zoom, and default gravity', () => {
         const cloudName = 'customtestcloud';
         const deliveryType = 'upload';
         const publicId = 'myimage';
-        const width = 900;
-        const widthResize = 900;
+        const aspectRatio = '16:9';
+        const crop = 'fill';
+        const zoom = '1.2';
 
         const url = constructCloudinaryUrl({
           options: {
             src: publicId,
-            width,
-            widthResize
+            aspectRatio,
+            crop,
+            zoom
           },
           config: {
             cloud: {
@@ -223,29 +227,8 @@ describe('Cloudinary', () => {
             }
           }
         });
-        expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/${deliveryType}/c_limit,w_${width}/c_limit,w_${widthResize}/f_auto/q_auto/${publicId}`);
-      });
 
-      it('should create a Cloudinary URL with widthResize larger than width', () => {
-        const cloudName = 'customtestcloud';
-        const deliveryType = 'upload';
-        const publicId = 'myimage';
-        const width = 900;
-        const widthResize = 1200;
-
-        const url = constructCloudinaryUrl({
-          options: {
-            src: publicId,
-            width,
-            widthResize
-          },
-          config: {
-            cloud: {
-              cloudName
-            }
-          }
-        });
-        expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/${deliveryType}/c_limit,w_${width}/c_limit,w_${widthResize}/f_auto/q_auto/${publicId}`);
+        expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/${deliveryType}/c_${crop},ar_${aspectRatio},g_auto,z_${zoom}/f_auto/q_auto/${publicId}`);
       });
 
     });
@@ -747,10 +730,12 @@ describe('Cloudinary', () => {
     it('Kitchen Sink - Image', () => {
       const cloudName = 'customtestcloud';
         const assetType = 'image';
+        const crop = 'auto';
+        const defaultImage = 'my-image.jpg';
+        const height = 321;
         const src = 'turtle';
         const width = 123;
-        const height = 321;
-        const defaultImage = 'my-image.jpg';
+        const zoom = '2.0';
 
         const cartoonify = '50';
         const gradientFade = true;
@@ -766,11 +751,27 @@ describe('Cloudinary', () => {
 
         const url = constructCloudinaryUrl({
           options: {
+            assetType,
+            crop: [
+              {
+                type: crop,
+                height,
+                width,
+                zoom,
+              },
+              {
+                type: crop,
+                height,
+                width,
+                zoom,
+                source: true
+              },
+            ],
+            defaultImage,
+            height,
             src,
             width,
-            height,
-            assetType,
-            defaultImage,
+            zoom,
             effects: [
               {
                 shear,
@@ -825,6 +826,7 @@ describe('Cloudinary', () => {
             }
           }
         });
+
         expect(url).toContain([
           assetType,
           `upload`,
@@ -832,13 +834,14 @@ describe('Cloudinary', () => {
           `e_gen_remove:prompt_apple;multiple_true;remove-shadow_true`,
           `e_background_removal`,
           `e_gen_restore`,
-          `c_limit,w_${width}`,
+          `c_${crop},w_${width},h_${height},g_auto,z_${zoom}`,
           `d_${defaultImage}`,
           `o_${opacity},e_shear:${shear}`,
           `e_cartoonify:${cartoonify},e_gradient_fade,r_${radius}`,
           `l_${overlaySrc},co_${overlayColor},e_shadow:${overlayShadow},x_${overlayX},y_${overlayY}`,
           `fl_layer_apply,fl_no_overflow,co_${overlayColor},e_shadow:${overlayShadow},x_${overlayX},y_${overlayY}`,
           `e_zoompan`,
+          `c_${crop},w_${width},h_${height},g_auto,z_${zoom}`,
           `f_auto:animated`, // Effect of zoompan
           `q_auto`,
           src